rename options to sign in card
